Support reading from record type

+++ src/cz/cuni/mff/d3s/trupple/language/parser/identifierstable/types/constant/CharConstantDescriptor.java
@@ -14,7 +14,7 @@ public class CharConstantDescriptor implements OrdinalConstantDescriptor {
 
     @Override
     public FrameSlotKind getSlotKind() {
-        return FrameSlotKind.Long;
+        return FrameSlotKind.Byte;
     }
 
     @Override
